refactor: constants 파일 inputs, questions로 분리
