+++ src/constants/inputs.ts
@@ -1,0 +1,3 @@
+export const MAX_NICKNAME_LENGTH = 15;
+export const MAX_PASSWORD_LENGTH = 15;
+export const MAX_COMMENT_LENGTH = 100;
