+++ src/constants/questions.ts
@@ -5,7 +5,3 @@ export const COUPLE = 'couple';
 export const FRIEND = 'friend';
 export const RANDOM = 'random';
 export const USERMADE = 'usermade';
-
-export const MAX_NICKNAME_LENGTH = 15;
-export const MAX_PASSWORD_LENGTH = 15;
-export const MAX_COMMENT_LENGTH = 100;
